fix missing fixtures
